requirements.txt
aiohttp==3.5.4
pytest==5.0.1
PyHamcrest==1.9.0
